requirements.txt
fastapi
uvicorn
pillow
pandas
transformers
torchvision
torch
python-multipart
